fix: disable context menu (right click) events
The context menu (right click menu) causes a usability issue for users
on devices where it is easy to accidentally trigger them. For instance
when using windows based touch devices it is vastly easier to
accidentally trigger a right click event which would trigger the context
menu. Given the nature of the application this is not desirable and
prevents the expected actions from happening.