Update EntityManager.ts
Ward should result in taking half damage rounded down. Since the code is written for healing the damage, I believe it should be 
entity.health += ceil(damage taken / 2), assuming entity.health already has incorporated the full damage. 

For brittle, the damage should be double what was done. It looks like the code was using the same function as the ward and subtracting, so the value was less than double. If I understand the code correctly, I think changing it to just condition.value should fix that?